[ 1816580 ] Make classes easier to add without breaking image. [ 1816583 ] Make it easier to add new classes to ooRexx [ 1837129 ] Add block synchronization objects
git-svn-id: https://svn.code.sf.net/p/oorexx/code-0/interpreter-3.x/trunk@1608 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ kernel/runtime/RexxEnvelope.hpp
@@ -50,7 +50,7 @@
                                        /* a translated method in EA's         */
 
 #define BEHAVIOUR_NON_PRIMITIVE 0x80000000
-#define ObjectHasNonPrimitiveBehaviour(o) (((long)(((RexxObject *)o)->behaviour)) & BEHAVIOUR_NON_PRIMITIVE)
+#define ObjectHasNonPrimitiveBehaviour(o) (((uintptr_t)(((RexxObject *)o)->behaviour)) & BEHAVIOUR_NON_PRIMITIVE)
 
 class RexxSmartBuffer;
